+++ services/madoc-ts/src/frontend/site/blocks/AllProjectsPaginatedItems.tsx
@@ -3,24 +3,29 @@ import React from 'react';
 import { blockEditorFor } from '../../../extensions/page-blocks/block-editor-for';
 import { SingleProject } from './SingleProject';
 import { useProjectList } from '../hooks/use-project-list';
+import { SingleProjectItem } from '../../tailwind/blocks/SingleProjectItem';
 
 interface AllProjectPaginatedItemsProps {
   customButtonLabel?: InternationalString;
   background?: string;
   radius?: string;
+  grid?: boolean;
 }
 
 export const AllProjectsPaginatedItems: React.FC<AllProjectPaginatedItemsProps> = ({
   customButtonLabel,
   radius,
   background,
+  grid,
 }) => {
   const { resolvedData: data } = useProjectList();
 
+  const Component = grid ? SingleProjectItem : SingleProject;
+
   return (
-    <>
+    <div style={grid ? { display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 20 } : undefined}>
       {data?.projects.map(project => (
-        <SingleProject
+        <Component
           key={project.id}
           project={{ id: project.slug }}
           data={project as any}
@@ -29,7 +34,7 @@ export const AllProjectsPaginatedItems: React.FC<AllProjectPaginatedItemsProps> 
           background={background}
         />
       ))}
-    </>
+    </div>
   );
 };
 
@@ -41,6 +46,7 @@ blockEditorFor(AllProjectsPaginatedItems, {
     customButtonLabel: '',
     background: null,
     radius: null,
+    grid: false,
   },
   source: {
     name: 'All projects page',
@@ -53,5 +59,6 @@ blockEditorFor(AllProjectsPaginatedItems, {
     customButtonLabel: { type: 'text-field', label: 'Custom button label' },
     background: { type: 'color-field', label: 'Background color', defaultValue: '#eeeeee' },
     radius: { type: 'text-field', label: 'Border radius', defaultValue: '' },
+    grid: { type: 'checkbox-field', label: 'Use new grid', inlineLabel: 'enable grid' },
   },
 });